+++ CHANGELOG.rst
@@ -35,6 +35,8 @@ Changelog
     - Renamed ``page`` to ``current_page``, ``results`` to ``results_count``, and ``fully_fetched`` to ``is_fully_fetched`` in ``ItemSummary``, ``Mounts``, ``Familiars`` and ``Outfits``.
     - Renamed ``CMPostArchive.page`` and ``CMPostArchive.posts`` to ``current_page`` and ``entries`` respectively.
     - Renamed ``Killer``to ``DeathParticipant``
+    - Renamed ``ForumBoard.threads`` to ``entries``.
+    - Added ``thread_starter_deleted`` to ``ThreadEntry``.
 - Added `ForumSection`` model and its respective parser, to fetch a list of board entries.
 - Removed ``get_url`` class methods from all models, replaced by functions in the urls package.
 - Fixed ``Character`` account badges not being parsed properly.